+++ tools/generate-resource-tpl.js
@@ -209,6 +209,14 @@ function schemaField(name, resourceSchema, requiredFields, pathIdField) {
 	const schema = resourceSchema.properties[name]
 	const optional = (requiredFields || []).indexOf(name) === -1 || schema.enum ? "true" : "false"
 	const required = (requiredFields || []).indexOf(name) === -1 || schema.enum ? "false" : "true"
+	let defaultValue;
+	if (schema.default) {
+		defaultValue = `Default: "${schema.default}"`
+	} else if (schema.enum && schema.enum.length > 0) {
+		defaultValue = `Default: "${schema.enum[0]}"`
+	} else {
+		defaultValue = `Computed: ${optional}`
+	}
 	const description = annotatedDescription(schema)
 	const forceNew = name === pathIdField || schema.write_only ? "true" : "false"
 	const skipDiff = schema.write_only ? `
@@ -227,7 +235,7 @@ function schemaField(name, resourceSchema, requiredFields, pathIdField) {
 			return `
 			"${name}": &schema.Schema {
 				Type: schema.TypeString,
-				${schema.enum ? `Default: "${schema.enum[0]}",` : `Computed: ${optional},`}
+				${defaultValue},
 				Required: ${required},
 				Optional: ${optional},
 				ForceNew: ${forceNew},